+++ example/lib/pages/home_page.dart
@@ -31,6 +31,7 @@ class HomePage extends StatelessWidget {
     'sheet_tabs': Icons.tab,
     'defined_names': Icons.label_outline,
     'rich_text': Icons.format_color_text,
+    'conditional_format': Icons.gradient,
   };
 
   static const _showcaseIcons = <String, IconData>{